requirements.txt
requests
beautifulsoup4
psycopg2-binary
flask
flask-cors
python-dotenv
pytz

